Refactor to use scene module
Simplifies a lot of the code and scene switching code.

The use of local functions in the gameplay scene is an attempt to make
the code more performant. But it's annoying that they have to be
declared above anywhere they're called. Not quite sure if that's ideal
or not yet.